+++ src/configmanager.h
@@ -138,7 +138,7 @@ class ConfigManager
 		bool reload();
 
 		const std::string& getString(string_config_t what) const;
-		int32_t getNumber(integer_config_t what) const;
+		int16_t getNumber(integer_config_t what) const;
 		bool getBoolean(boolean_config_t what) const;
 		float getFloat(floating_config_t what) const;
 
@@ -153,7 +153,7 @@ class ConfigManager
 	private:
 		std::string configFileLua = { "config.lua" };
 		std::string string[LAST_STRING_CONFIG] = {};
-		int32_t integer[LAST_INTEGER_CONFIG] = {};
+		int16_t integer[LAST_INTEGER_CONFIG] = {};
 		bool boolean[LAST_BOOLEAN_CONFIG] = {};
 		float floating[LAST_FLOATING_CONFIG] = {};
