+++ src/tile.h
@@ -146,7 +146,7 @@ class TileItemVector : private ItemVector
 		}
 
 	private:
-		uint16_t downItemCount = 0;
+		uint32_t downItemCount = 0;
 };
 
 class Tile : public Cylinder
@@ -265,7 +265,6 @@ class Tile : public Cylinder
 
 		void postAddNotification(Thing* thing, const Cylinder* oldParent, int32_t index, cylinderlink_t link = LINK_OWNER) override final;
 		void postRemoveNotification(Thing* thing, const Cylinder* newParent, int32_t index, cylinderlink_t link = LINK_OWNER) override final;
-
 		void internalAddThing(Thing* thing) override final;
 		void internalAddThing(uint32_t index, Thing* thing) override;
